report: copy ContextPath when adding it to a report.Entry
ContextPath.Append() makes a new ContextPath struct and path slice, but
the slice sometimes refers to the original underlying array, depending
what append() decides to do.  As a result, anything wanting to persist a
path outside the current call chain needs to make a Copy() and persist
that.  Otherwise the following pattern would be unsafe:

    r.AddOnError(c.Append("mode"), validateMode(f.Mode))
    r.AddOnError(c.Append("overwrite"), f.validateOverwrite())

Report was failing to do this, and it seemed fine on amd64, but on 386
it caused one Butane test to fail.  Fix Report to copy paths.  Ignition
merge.Transcript and Butane translate.TranslationSet already do this.

Alternatively, we could make Append() always copy the path slice.  But
there are a lot of call sites which build paths and then throw them away,
and very few places that actually persist paths, so it seems feasible and
more efficient to retain the current Append() behavior for now.